requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
Werkzeug==2.2.2
tensorflow==2.10.0
pillow==9.3.0
pandas==1.5.3
numpy==1.23.3
secrets==1.0
matplotlib==3.6.2
